Code refactored. All warnings fixed. Link to specyfication of XPM format added.
